+++ agent/src/main/java/com/intergral/deep/agent/push/PushService.java
@@ -20,7 +20,6 @@ package com.intergral.deep.agent.push;
 import com.intergral.deep.agent.api.plugin.ISnapshotContext;
 import com.intergral.deep.agent.api.plugin.ISnapshotDecorator;
 import com.intergral.deep.agent.api.resource.Resource;
-import com.intergral.deep.agent.api.spi.IDeepPlugin;
 import com.intergral.deep.agent.grpc.GrpcService;
 import com.intergral.deep.agent.settings.Settings;
 import com.intergral.deep.agent.types.snapshot.EventSnapshot;
@@ -59,17 +58,15 @@ public class PushService {
   }
 
   private void decorate(final EventSnapshot snapshot, final ISnapshotContext context) {
-    final Collection<IDeepPlugin> plugins = this.settings.getPlugins();
-    for (IDeepPlugin plugin : plugins) {
-      if (plugin instanceof ISnapshotDecorator) {
-        try {
-          final Resource decorate = ((ISnapshotDecorator) plugin).decorate(this.settings, context);
-          if (decorate != null) {
-            snapshot.mergeAttributes(decorate);
-          }
-        } catch (Throwable t) {
-          LOGGER.error("Error processing plugin {}", plugin.getClass().getName());
+    final Collection<ISnapshotDecorator> plugins = this.settings.getPlugins(ISnapshotDecorator.class);
+    for (ISnapshotDecorator plugin : plugins) {
+      try {
+        final Resource decorate = plugin.decorate(this.settings, context);
+        if (decorate != null) {
+          snapshot.mergeAttributes(decorate);
         }
+      } catch (Throwable t) {
+        LOGGER.error("Error processing plugin {}", plugin.getClass().getName());
       }
     }
     snapshot.close();