+++ agent/src/main/java/com/intergral/deep/agent/resource/SpiUtil.java
@@ -5,13 +5,11 @@
 
 package com.intergral.deep.agent.resource;
 
-import com.intergral.deep.agent.api.spi.ConditionalProvider;
 import com.intergral.deep.agent.api.spi.Ordered;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.ServiceLoader;
-import java.util.stream.Collectors;
 
 /**
  * Utilities to load SPI services.
@@ -24,16 +22,6 @@ public final class SpiUtil {
   public static <T extends Ordered> List<T> loadOrdered(Class<T> spiClass,
       ClassLoader serviceClassLoader) {
     return loadOrdered(spiClass, serviceClassLoader, ServiceLoader::load);
-  }
-
-  public static <T extends Ordered> List<T> loadConditional(Class<T> spiClass, ClassLoader serviceClassLoader) {
-    final List<T> ordered = loadOrdered(spiClass, serviceClassLoader, ServiceLoader::load);
-    return ordered.stream().filter(s -> {
-      if (s instanceof ConditionalProvider) {
-        return ((ConditionalProvider) s).isActive();
-      }
-      return true;
-    }).collect(Collectors.toList());
   }
 
   // VisibleForTesting