+++ agent/src/test/java/com/intergral/deep/test/MockEventSnapshot.java
@@ -31,8 +31,8 @@ import java.util.HashSet;
 public class MockEventSnapshot extends EventSnapshot {
 
   public MockEventSnapshot() {
-    super(new TracePointConfig("tp-1", "/some/file/path.py", 123, new HashMap<>(), new ArrayList<>(), new ArrayList<>()), 1011L, Resource.DEFAULT,
-        new ArrayList<>(), new HashMap<>());
+    super(new TracePointConfig("tp-1", "/some/file/path.py", 123, new HashMap<>(), new ArrayList<>(), new ArrayList<>()), 1011L,
+        Resource.DEFAULT, new ArrayList<>(), new HashMap<>());
   }
 
   public MockEventSnapshot withFrames() {